backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
cryptography==44.0.1
python-dotenv==1.0.1
netmiko==4.5.0
napalm==5.0.0
napalm-huawei-vrp==1.2.0
pysnmp==7.1.18
requests==2.32.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
